[DIDA] optimizing import , remove unused Resource 수행 , Id관련 자료형 통일

+++ common/src/main/java/com/dida/common/actionhandler/NftActionHandler.kt
@@ -1,6 +1,6 @@
 package com.dida.common.actionhandler
 
 interface NftActionHandler {
-    fun onNftItemClicked(nftId: Int)
-    fun onLikeBtnClicked(nftId: Int)
+    fun onNftItemClicked(nftId: Long)
+    fun onLikeBtnClicked(nftId: Long)
 }
