+++ domain/src/main/java/com/dida/domain/model/nav/hide/CardHideListItem.kt
@@ -1,7 +1,7 @@
 package com.dida.domain.model.nav.hide
 
 data class CardHideList(
-    val cardId: Int,
+    val cardId: Long,
     val cardUrl: String,
     val cardTitle: String,
     val userName : String,
